chore: add more german translations (#1153)
Hi,
mapunik works great for editing styles. Thanks for the tooling ❤️

When launching mapuntik in the german version, there a bunch of warnings
in the log that come from untranslated strings.

This fixes a few of them and also runs the translation script.
The translation script is unable to find the strings to be translated,
unsure how this works.


![image](https://github.com/user-attachments/assets/cbc8614d-f3c2-4519-bec1-a5dd5c573d5f)



## Launch Checklist

<!-- Thanks for the PR! Feel free to add or remove items from the
checklist. -->


 - [x] Briefly describe the changes in this PR.
 - [ ] Link to related issues.
- [x] Include before/after visuals or gifs if this PR includes visual
changes.
 - [ ] Write tests for all new functionality.
 - [ ] Add an entry to `CHANGELOG.md` under the `## main` section.

---------

Co-authored-by: Harel M <harel.mazor@gmail.com>